refactor(worker): extract env-var parsing into config.go
Pull workerConfig and loadConfig into a new cmd/worker/config.go so
main.go becomes wiring only. Operators can now answer "what env vars
does the worker read?" by grepping a single file. The dense workflow-
tuning block (MIN_CROP_X/Y, DETECT_CROP/TRANSCODE/PROGRESS_LOG_INTERVAL
timeouts, H265_CRF, HARDWARE_DEVICE_PATH, METRICS_HIGH_CARDINALITY_LABELS)
extracts further into loadWorkflowConfig for isolated readability and
testing. validateHardwareDevicePath moves alongside the other env helpers
and is invoked from loadWorkflowConfig so all env-var validation surfaces
at the same point.

Helper tests move from run_test.go to config_test.go (file renamed via
git mv to preserve history). TestRun_MissingTaskQueue becomes
TestLoadConfigMissingTaskQueue and calls loadConfig directly, since the
TEMPORAL_TASK_QUEUE check lives in the config layer now.

METRICS_ADDR continues to be read inside metrics.NewFromEnv. No behavior
change beyond the call-site move of validateHardwareDevicePath.

Fixes #187
